Add files via upload
The excel file is too big to upload here so I only uploaded a text file of the VBA code.

The bonus section was a bit more challenging. I tried several code snippets but I couldn't get one that executes the task as required. I submitted the best code I got, even though it is obvious that it did not do the job.